Extract stuff from Program.cs
